+++ NEWS.md
@@ -1,4 +1,5 @@
 # themask 1.0.0
+* Address bug with  `msk_available` that was yielding an error when run [2023-10-10]
 * Built in connection to Google Sheet to maintain mechanisms across datasets and periods [2023-10-10]
 * Adjust `minoria_geo`, removing suffix from names (`_milb`) [2023-10-03]
 * Add in shape files to the package (and as releases) for use in mapping [2023-10-03]